# MedicalDashboard/OpenAPI-documentation.yaml
openapi: 3.0.3
info:
  title: Medical Dashboard API
  description: |    
    ### Микросервисы:
    - **AuthService** - Аутентификация и управление пользователями
    - **DashboardAPI** - Управление пациентами, метриками и алертами
    - **NotificationService** - Отправка уведомлений
    
    ### Аутентификация
    Большинство эндпоинтов требуют JWT Bearer токен через `/identity/login`.
    
    ### SignalR
    Для real-time получения метрик используется SignalR Hub `/hubs/metrics` через Gateway.
    Gateway автоматически проксирует WebSocket соединения к DashboardAPI.
    Поддерживаются методы:
    - `SubscribeToPatient(patientId: Guid)` - подписка на метрики пациента
    - `SubscribeToPatientMetrics(patientId: Guid)` - подписка на метрики пациента (альтернативное имя)
    - `UnsubscribeFromPatient(patientId: Guid)` - отписка от метрик пациента
    
    ### Ограничения
    - Все запросы должны содержать заголовок `Content-Type: application/json`
    - JWT токены имеют ограниченное время жизни - 7 дней
  version: 1.0.0
servers:
  - url: http://localhost:7168
    description: Development Gateway (проксирует запросы к микросервисам)
security:
  - BearerAuth: []
paths:
...
  /dead-letters:
    get:
      tags:
        - DeadLetter
      summary: Получить все сообщения Dead Letter
      description: Получение списка всех сообщений в очереди Dead Letter
      operationId: getAllDeadLetters
      responses:
        '200':
          description: Список сообщений Dead Letter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeadLetterMessage'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResult'
  /dead-letters/unprocessed:
    get:
      tags:
        - DeadLetter
      summary: Получить необработанные сообщения Dead Letter
      description: Получение списка необработанных сообщений в очереди Dead Letter
      operationId: getUnprocessedDeadLetters
      responses:
        '200':
          description: Список необработанных сообщений Dead Letter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeadLetterMessage'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResult'
  /dead-letters/{id}/process:
    post:
      tags:
        - DeadLetter
      summary: Отметить сообщение Dead Letter как обработанное
      description: Отметить сообщение в очереди Dead Letter как обработанное
      operationId: processDeadLetter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Идентификатор сообщения Dead Letter
      responses:
        '200':
          description: Сообщение успешно обработано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadLetterMessage'
        '404':
          description: Сообщение не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResult'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResult'

  # ========================================
  # DEVICES ENDPOINTS
  # ========================================
  /devices:
    get:
      tags: [Devices]
      summary: Получить список устройств
      description: Получение списка устройств с фильтрацией по палате и статусу inUsing
      security: [BearerAuth: []]
      parameters:
        - name: ward
          in: query
          schema:
            type: integer
          description: Фильтр по номеру палаты
        - name: inUsing
          in: query
          schema:
            type: boolean
          description: Фильтр по статусу занятости
      responses:
        '200':
          description: Список устройств
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceDto'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Devices]
      summary: Создать устройство
      description: Создание нового устройства
      security: [BearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiDeviceDto'
      responses:
        '201':
          description: Устройство создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDto'
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /devices/{id}:
    get:
      tags: [Devices]
      summary: Получить устройство по ID
      description: Получение информации об устройстве по его ID
      security: [BearerAuth: []]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID устройства
      responses:
        '200':
          description: Информация об устройстве
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDto'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Devices]
      summary: Обновить устройство
      description: Обновление информации об устройстве
      security: [BearerAuth: []]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID устройства
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiDeviceDto'
      responses:
        '200':
          description: Устройство обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDto'
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Devices]
      summary: Удалить устройство
      description: Удаление устройства по ID
      security: [BearerAuth: []]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID устройства
      responses:
        '204':
          description: Устройство удалено
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /devices/on-patient/{id}:
    get:
      tags: [Devices]
      summary: Получить устройства, подключённые к пациенту
      description: Получение краткой информации об устройствах, подключённых к пациенту
      security: [BearerAuth: []]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID пациента
      responses:
        '200':
          description: Список устройств и уникальные метрики
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceShortDto'
                  metrics:
                    type: array
                    items:
                      type: string
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пациент не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /devices/attach:
    post:
      tags: [Devices]
      summary: Привязать устройство к пациенту
      description: Привязка устройства к пациенту (оба id в теле)
      security: [BearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachDeviceDto'
      responses:
        '200':
          description: Устройство привязано
        '400':
          description: Неверные данные или палаты не совпадают
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Устройство или пациент не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /devices/{id}/detach:
    post:
      tags: [Devices]
      summary: Отвязать устройство от пациента
      description: Отвязка устройства от пациента
      security: [BearerAuth: []]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID устройства
      responses:
        '200':
          description: Устройство отвязано
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer токен для аутентификации
  schemas:
...
    NotificationResult:
      type: object
      properties:
        success:
          type: boolean
          description: Успешность операции
          example: true
        errorMessage:
          type: string
          nullable: true
          description: Сообщение об ошибке
          example: "Failed to send notification"

    # ========================================
    # COMMON SCHEMAS
    # ========================================
    ErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          description: Детальное сообщение об ошибке
          example: "Detailed error message"
    DeadLetterMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор сообщения Dead Letter
          example: "123e4567-e89b-12d3-a456-426614174000"
        messageId:
          type: string
          format: uuid
          description: Уникальный идентификатор сообщения в очереди
          example: "123e4567-e89b-12d3-a456-426614174001"
        body:
          type: string
          description: Тело сообщения
          example: "{\"patientId\":\"123e4567-e89b-12d3-a456-426614174000\",\"type\":\"heart_rate\",\"value\":75.0,\"timestamp\":\"2024-01-15T10:30:00Z\"}"
        headers:
          type: object
          additionalProperties:
            type: string
          description: Заголовки сообщения
          example:
            Content-Type: "application/json"
            X-Correlation-Id: "1234567890"
        deliveryCount:
          type: integer
          description: Количество попыток доставки
          example: 1
        enqueuedAt:
          type: string
          format: date-time
          description: Время поступления в очередь
          example: "2024-01-15T10:30:00Z"
        lockedUntil:
          type: string
          format: date-time
          nullable: true
          description: Время разблокировки сообщения
          example: "2024-01-15T10:30:00Z"
        deadLetterReason:
          type: string
          description: Причина Dead Letter
          example: "MaxDeliveryCountReached"
        errorMessage:
          type: string
          nullable: true
          description: Сообщение об ошибке при доставке
          example: "Failed to process message"
        lastDeliveryAttempt:
          type: string
          format: date-time
          nullable: true
          description: Время последней попытки доставки
          example: "2024-01-15T10:30:00Z"

    # ========================================
    # DEVICES SCHEMAS
    # ========================================
    DeviceDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор устройства
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Название устройства
          example: "Pulse-meter 1"
        ward:
          type: integer
          description: Номер палаты, к которой привязано устройство
          example: 101
        inUsing:
          type: boolean
          description: Статус устройства
          example: "true"
        readableMetrics:
          type: array
          description: Считываемые устройством метрики
          example: ["Pulse"]
        patientId:
          type: string
          format: uuid
          nullable: true
          description: ID пациента, к которому привязано устройство (если в использовании)
          example: "123e4567-e89b-12d3-a456-426614174001"
    ApiDeviceDto:
      type: object
      required:
        - name
        - ward
        - readableMetrics
      properties:
        name:
          type: string
          description: Название устройства
          example: "Pulse-meter 1"
        ward:
          type: integer
          description: Номер палаты, к которой привязать устройство
          example: 101
        readableMetrics:
          type: array
          description: Считываемые метрики
          example: ["Pulse"]
    DeviceShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор устройства
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Название устройства
          example: "Pulse-meter 1"
        readableMetrics:
          type: array
          description: Считываемые метрики
          example: [ "Pulse" ]
    AttachDeviceDto:
      type: object
      required:
        - deviceId
        - patientId
      properties:
        deviceId:
          type: string
          format: uuid
          description: ID устройства для привязки
          example: "123e4567-e89b-12d3-a456-426614174000"
        patientId:
          type: string
          format: uuid
          description: ID пациента для привязки
          example: "123e4567-e89b-12d3-a456-426614174001"
...